+++ lib/auth0_ex/token_provider/encrypted_redis_token_cache.ex
@@ -29,7 +29,7 @@ defmodule Auth0Ex.TokenProvider.EncryptedRedisTokenCache do
         {:ok, nil}
 
       {:ok, cached_value} ->
-        parse_and_decrypt(cached_value)
+        decrypt_and_parse(cached_value)
 
       {:error, reason} ->
         Logger.warn("Error retrieving token from redis.", audience: audience, key: key, reason: reason)
@@ -59,7 +59,7 @@ defmodule Auth0Ex.TokenProvider.EncryptedRedisTokenCache do
     end
   end
 
-  defp parse_and_decrypt(cached_value) do
+  defp decrypt_and_parse(cached_value) do
     with {:ok, decrypted} <- TokenEncryptor.decrypt(cached_value),
          {:ok, token_attributes} <- Jason.decode(decrypted) do
       build_token(token_attributes)
@@ -72,8 +72,8 @@ defmodule Auth0Ex.TokenProvider.EncryptedRedisTokenCache do
 
   defp to_json(token), do: Jason.encode(token)
 
-  defp build_token(%{"jwt" => jwt, "issued_at" => issued_at, "expires_at" => expires_at}) do
-    {:ok, %TokenInfo{jwt: jwt, issued_at: issued_at, expires_at: expires_at}}
+  defp build_token(%{"jwt" => jwt, "issued_at" => issued_at, "expires_at" => expires_at} = token) do
+    {:ok, %TokenInfo{jwt: jwt, issued_at: issued_at, expires_at: expires_at, kid: token["kid"]}}
   end
 
   defp build_token(_), do: {:error, :malformed_cached_data}