+++ test/auth0_ex/token_provider/auth0_authorization_service_test.exs
@@ -53,6 +53,9 @@ defmodule Auth0Ex.TokenProvider.Auth0AuthorizationServiceTest do
     {:error, :request_error} = Auth0AuthorizationService.retrieve_token(credentials, "audience")
   end
 
-  defp sample_token, do: JwtUtils.jwt_that_expires_in(86_400, @test_audience)
+  defp sample_token do
+    JwtUtils.generate_fake_jwt(@test_audience, %{}, %{"kid" => "my-kid"})
+  end
+
   defp valid_auth0_response, do: ~s<{"access_token":"#{sample_token()}","expires_in":86400,"token_type":"Bearer"}>
 end