+++ test/auth0_ex/token_provider/token_info_test.exs
@@ -7,10 +7,11 @@ defmodule Auth0Ex.TokenProvider.TokenInfoTest do
   test "extracts metadata from token" do
     issued_at = Timex.now() |> Timex.shift(hours: -12) |> Timex.to_unix()
     expires_at = Timex.now() |> Timex.shift(hours: 12) |> Timex.to_unix()
-    token = JwtUtils.jwt_with_claims(%{iat: issued_at, exp: expires_at})
+    token = JwtUtils.generate_fake_jwt("some-audience", %{iat: issued_at, exp: expires_at}, %{kid: "my-kid"})
 
     assert %TokenInfo{
              jwt: token,
+             kid: "my-kid",
              issued_at: issued_at,
              expires_at: expires_at
            } == TokenInfo.from_jwt(token)