+++ test/auth0_ex/token_provider_test.exs
@@ -6,8 +6,8 @@ defmodule Auth0Ex.TokenProviderTest do
   alias Auth0Ex.TokenProvider.TokenInfo
 
   @sample_credentials %Auth0Credentials{base_url: "base_url", client_id: "client_id", client_secret: "client_secret"}
-  @sample_token %TokenInfo{jwt: "SAMPLE-TOKEN", issued_at: 123, expires_at: 234}
-  @another_sample_token %TokenInfo{jwt: "ANOTHER-SAMPLE-TOKEN", issued_at: 123, expires_at: 234}
+  @sample_token %TokenInfo{jwt: "SAMPLE-TOKEN", issued_at: 123, expires_at: 234, kid: "valid-kid"}
+  @another_sample_token %TokenInfo{jwt: "ANOTHER-SAMPLE-TOKEN", issued_at: 123, expires_at: 234, kid: "valid-kid"}
 
   setup :set_mox_global
   setup :verify_on_exit!
@@ -17,6 +17,7 @@ defmodule Auth0Ex.TokenProviderTest do
 
     in_one_hour = Timex.shift(Timex.now(), hours: 1)
     stub(RefreshStrategyMock, :refresh_time_for, fn _ -> in_one_hour end)
+    stub(JwksKidsFetcherMock, :fetch_kids, fn _ -> {:ok, ["valid-kid"]} end)
 
     {:ok, %{pid: pid}}
   end
@@ -48,7 +49,7 @@ defmodule Auth0Ex.TokenProviderTest do
 
     initialize_for_audience("target_audience", @sample_token, pid)
 
-    expect(TokenServiceMock, :refresh_token, 0, fn _, _, _ -> :should_never_be_called end)
+    stub(TokenServiceMock, :refresh_token, fn _, _, _ -> {:ok, @another_sample_token} end)
 
     wait_for_first_check_to_complete()
 
@@ -72,6 +73,25 @@ defmodule Auth0Ex.TokenProviderTest do
     assert {:ok, "ANOTHER-SAMPLE-TOKEN"} == TokenProvider.token_for(pid, "target_audience")
   end
 
+  test "refreshes its tokens when their signature is not valid (eg. for keys revoked)", %{pid: pid} do
+    sometime_after_next_check = Timex.shift(Timex.now(), hours: 1)
+    expect(RefreshStrategyMock, :refresh_time_for, 2, fn _ -> sometime_after_next_check end)
+
+    initialize_for_audience("target_audience", @sample_token, pid)
+
+    expect(JwksKidsFetcherMock, :fetch_kids, fn _credentials -> {:ok, ["different-kids"]} end)
+
+    expect(
+      TokenServiceMock,
+      :refresh_token,
+      fn @sample_credentials, "target_audience", @sample_token -> {:ok, @another_sample_token} end
+    )
+
+    wait_for_first_signature_check_to_complete()
+
+    assert {:ok, "ANOTHER-SAMPLE-TOKEN"} == TokenProvider.token_for(pid, "target_audience")
+  end
+
   defp initialize_for_audience(audience, token, pid) do
     expect(TokenServiceMock, :retrieve_token, fn %Auth0Credentials{}, ^audience -> {:ok, token} end)
 
@@ -79,6 +99,8 @@ defmodule Auth0Ex.TokenProviderTest do
   end
 
   defp wait_for_first_check_to_complete, do: :timer.sleep(token_check_interval() + 500)
-
   defp token_check_interval, do: Application.fetch_env!(:auth0_ex, :client)[:token_check_interval]
+
+  defp wait_for_first_signature_check_to_complete, do: :timer.sleep(signature_check_interval() + 500)
+  defp signature_check_interval, do: Application.fetch_env!(:auth0_ex, :client)[:signature_check_interval]
 end