+++ test/integration/encrypted_redis_token_cache_test.exs
@@ -1,4 +1,4 @@
-defmodule Auth0Ex.TokenProvider.EncryptedRedisTokenCacheTest do
+defmodule Integration.TokenProvider.EncryptedRedisTokenCacheTest do
   use ExUnit.Case, async: true
 
   import Auth0Ex.TestSupport.TimeUtils
@@ -19,6 +19,16 @@ defmodule Auth0Ex.TokenProvider.EncryptedRedisTokenCacheTest do
     :ok = EncryptedRedisTokenCache.set_token_for(@test_audience, token)
 
     assert {:ok, token} == EncryptedRedisTokenCache.get_token_for(@test_audience)
+  end
+
+  test "retrieves tokens set by a previous version of auth0_ex, hence without kid" do
+    issued_at = one_hour_ago()
+    expires_at = in_one_hour()
+    token_without_kid = %{jwt: "my-token", issued_at: issued_at, expires_at: expires_at}
+    :ok = EncryptedRedisTokenCache.set_token_for(@test_audience, token_without_kid)
+
+    assert {:ok, %TokenInfo{jwt: "my-token", issued_at: ^issued_at, expires_at: ^expires_at, kid: nil}} =
+             EncryptedRedisTokenCache.get_token_for(@test_audience)
   end
 
   test "returns {:ok, nil} when token is not cached" do
@@ -51,7 +61,10 @@ defmodule Auth0Ex.TokenProvider.EncryptedRedisTokenCacheTest do
     assert {:ok, nil} = EncryptedRedisTokenCache.get_token_for(@test_audience)
   end
 
-  defp sample_token, do: %TokenInfo{jwt: "my-token", issued_at: one_hour_ago(), expires_at: in_one_hour()}
+  defp sample_token do
+    %TokenInfo{jwt: "my-token", issued_at: one_hour_ago(), expires_at: in_one_hour(), kid: "my-kid"}
+  end
+
   defp token_key(audience), do: "auth0ex_tokens:#{namespace()}:#{audience}"
   defp in_two_seconds, do: Timex.now() |> Timex.shift(seconds: 2) |> Timex.to_unix()
   defp namespace, do: Application.fetch_env!(:auth0_ex, :client)[:cache_namespace]